Update to get topicLength as int instead of byte

+++ src/main/java/com/github/jmatss/send/packet/PublishPacket.java
@@ -2,12 +2,12 @@ package com.github.jmatss.send.packet;
 
 public class PublishPacket {
     public final int subMessageType;
-    public final byte topicLength;
+    public final int topicLength;
     public final String topic;
     public final int port;
     public final byte[] id;
 
-    public PublishPacket(int subMessageType, byte topicLength, String topic, int port, byte[] id) {
+    public PublishPacket(int subMessageType, int topicLength, String topic, int port, byte[] id) {
         this.subMessageType = subMessageType;
         this.topicLength = topicLength;
         this.topic = topic;
